refactor: update names of product-service and tidy up project
