+++ lib/product-service/getProductList/handler.ts
@@ -1,5 +1,4 @@
-import { Context } from "aws-lambda";
-import IAvailableProduct from "../../ProductService/IAvailableProduct";
+import IAvailableProduct from "../types/IAvailableProduct";
 
 const mockProducts: Record<string, IAvailableProduct> = {
   "1": {
@@ -25,10 +24,12 @@ const mockProducts: Record<string, IAvailableProduct> = {
   },
 };
 
-async function getProductById(id: string) {
-  return mockProducts[id] || null;
+export async function getProductsList() {
+  return Object.values(mockProducts);
 }
 
-export async function handler(event: any, context: Context) {
-  return getProductById(event.id);
+export async function handler(event: any) {
+  const products = await getProductsList();
+
+  return products;
 }
