+++ lib/product-service/product-service-stack.ts
@@ -3,9 +3,9 @@ import { Construct } from "constructs";
 import createApi from "./stack/api";
 import createGetProductList from "./getProductList";
 import { WHITELISTED_ORIGINS } from "../constants";
-import createGetProductsById from "./getProductById";
+import createGetProductById from "./getProductById";
 
-export class ProductLambdaStack extends cdk.Stack {
+export class ProductServiceStack extends cdk.Stack {
   constructor(scope: Construct, id: string, props?: cdk.StackProps) {
     super(scope, id, props);
 
@@ -32,7 +32,7 @@ export class ProductLambdaStack extends cdk.Stack {
       ],
     });
 
-    const getProductById = createGetProductsById(this);
+    const getProductById = createGetProductById(this);
 
     const productIdResource = productResource.addResource("{id}");
