+++ lib/product-service/types/IAvailableProduct.ts
