+++ lib/product-service/types/IProduct.ts
